Construindo a tela de histórico de jogos do usuário

+++ app/src/main/java/com/example/teste_dynamox/src/activities/telas/TelaDeResultado.kt
@@ -55,7 +55,7 @@ fun TelaDeResultado(navController: NavController) {
         }
     }
 
-    fun inserirJogoPeloUserIdEReiniciarOQuiz(userId: Long?, quantDeAcertos: Int, quantDeErros: Int) {
+    fun inserirJogoPeloUserIdEReiniciarOQuiz(userId: Long?, quantDeAcertos: Long, quantDeErros: Long) {
         CoroutineScope(Dispatchers.IO).launch {
             jogosDao.inserirJogo(
                 jogosDosUsuaios(
@@ -75,7 +75,7 @@ fun TelaDeResultado(navController: NavController) {
         }
     }
 
-    fun inserirJogoPeloUserIdEMostrarHistorico(userId: Long?, quantDeAcertos: Int, quantDeErros: Int) {
+    fun inserirJogoPeloUserIdEMostrarHistorico(userId: Long?, quantDeAcertos: Long, quantDeErros: Long) {
         CoroutineScope(Dispatchers.IO).launch {
             jogosDao.inserirJogo(
                 jogosDosUsuaios(
